# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Southbank Flood Early-Warning Prototype",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/sensors/readings": {
      "post": {
        "summary": "Sensor Ingest",
        "operationId": "sensor_ingest_api_v1_sensors_readings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SensorReading"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/scenarios/{scenario_id}/start": {
      "post": {
        "summary": "Scenario Start",
        "operationId": "scenario_start_api_v1_scenarios__scenario_id__start_post",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/scenarios/manual": {
      "post": {
        "summary": "Scenario Manual",
        "operationId": "scenario_manual_api_v1_scenarios_manual_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualControl"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/settings": {
      "post": {
        "summary": "Settings",
        "operationId": "settings_api_v1_settings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/telemetry/render": {
      "post": {
        "summary": "Telemetry",
        "operationId": "telemetry_api_v1_telemetry_render_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenderTelemetry"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ManualControl": {
        "properties": {
          "level": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Level"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "level"
        ],
        "title": "ManualControl"
      },
      "RenderTelemetry": {
        "properties": {
          "readingId": {
            "type": "string",
            "format": "uuid",
            "title": "Readingid"
          },
          "renderedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Renderedat"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "readingId",
          "renderedAt"
        ],
        "title": "RenderTelemetry"
      },
      "SensorReading": {
        "properties": {
          "readingId": {
            "type": "string",
            "format": "uuid",
            "title": "Readingid"
          },
          "sensorId": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "pattern": "^[a-zA-Z0-9_.:-]+$",
            "title": "Sensorid"
          },
          "siteId": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "pattern": "^[a-zA-Z0-9_.:-]+$",
            "title": "Siteid"
          },
          "observedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Observedat"
          },
          "receivedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Receivedat"
          },
          "scenarioLevel": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Scenariolevel"
          },
          "waterDepthMm": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 10000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Waterdepthmm"
          },
          "rainfallIntensityMmPerHour": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Rainfallintensitymmperhour"
          },
          "trend": {
            "type": "string",
            "enum": [
              "rising",
              "falling",
              "steady",
              "unknown"
            ],
            "title": "Trend",
            "default": "unknown"
          },
          "trendPerMinute": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trendperminute"
          },
          "batteryPercent": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Batterypercent"
          },
          "quality": {
            "type": "string",
            "enum": [
              "valid",
              "stale",
              "fault",
              "unknown"
            ],
            "title": "Quality",
            "default": "valid"
          },
          "source": {
            "type": "string",
            "enum": [
              "mock",
              "physical"
            ],
            "title": "Source"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "readingId",
          "sensorId",
          "siteId",
          "observedAt",
          "scenarioLevel",
          "source"
        ],
        "title": "SensorReading",
        "description": "Stable ingestion contract. receivedAt is overwritten by the server."
      },
      "SettingsUpdate": {
        "properties": {
          "mode": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "simulation",
                  "normal"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Mode"
          },
          "impactScope": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "local_segment",
                  "full_route_demo"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Impactscope"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "SettingsUpdate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}